LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/blockeduri-eval-expected.txt


FAIL Eval violations have a blockedURI of 'eval' assert_equals: expected 12 but got 17
